requirement.txt
langchain
python-dotenv
langchain-community
langchain-groq
streamlit